requirements.txt
telebot~=0.0.5
pandas~=1.3.5
matplotlib~=3.5.3
Pillow~=9.5.0